Linux platform requires atomic instructions
- remove atomic_shim.h and rename atomic_a.h to atomic_arch.h
